Student2-Matej is now MAIN

+++ src/main/java/cz/uhk/pro2/tulipani/domain/entity/GroupRole.java
@@ -24,7 +24,7 @@ public class GroupRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Long roleId;
+    private Integer roleId;
 
     @Column(name = "role_name")
     private String roleName;
