+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package cz.uhk.pro2.tulipani.domain.repository;
 import cz.uhk.pro2.tulipani.domain.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
