+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/GroupRoleRepository.java
@@ -3,5 +3,5 @@ package cz.uhk.pro2.tulipani.domain.repository;
 import cz.uhk.pro2.tulipani.domain.entity.GroupRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRoleRepository extends JpaRepository<GroupRole, Long> {
+public interface GroupRoleRepository extends JpaRepository<GroupRole, Integer> {
 }
