+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/TaskCommentRepository.java
@@ -3,5 +3,5 @@ package cz.uhk.pro2.tulipani.domain.repository;
 import cz.uhk.pro2.tulipani.domain.entity.TaskComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskCommentRepository extends JpaRepository<TaskComment, Long> {
+public interface TaskCommentRepository extends JpaRepository<TaskComment, Integer> {
 }
