+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/UserSettingsRepository.java
@@ -3,5 +3,5 @@ package cz.uhk.pro2.tulipani.domain.repository;
 import cz.uhk.pro2.tulipani.domain.entity.UserSettings;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserSettingsRepository extends JpaRepository<UserSettings, Long> {
+public interface UserSettingsRepository extends JpaRepository<UserSettings, Integer> {
 }
